flask_sistema/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
# PostgreSQL (para produção futura)
# psycopg2-binary==2.9.7
python-dotenv==1.0.0
pandas==2.0.3
geopy==2.3.0
folium==0.14.0
plotly==5.15.0
gunicorn==21.2.0
